feat: [STEP3] 상품 등록 화면 구현

+++ OpenMarket/OpenMarket/OpenMarketAPI/RequestType/OpenMarketProductPostParm.swift
@@ -10,9 +10,9 @@ import Foundation
 struct OpenMarketProductPostParam: Codable {
     var name: String
     var descriptions: String
-    var price: Int
+    var price: Float
     var currency: OpenMarketProduct.Currency
-    var discountedPrice: Int = 0
+    var discountedPrice: Float = 0.0
     var stock: Int = 0
     var secret: String
